bank_ai_demo/requirements.txt
# Glacier First Bank AI Intelligence Demo - Python Dependencies

# Core Snowflake dependencies
snowflake-snowpark-python>=1.11.1
snowflake-connector-python>=3.6.0

# Configuration and data handling
PyYAML>=6.0.1
pandas>=2.0.0
numpy>=1.24.0

# Logging and utilities
python-dateutil>=2.8.2
uuid>=1.30

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
